requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.35
sqlalchemy[asyncio]>=2.0.35
alembic>=1.14.0
pydantic>=2.10.0
pydantic[email]>=2.10.0
pydantic-settings>=2.7.0
# PostgreSQL with PostGIS support
psycopg2-binary>=2.9.9
geoalchemy2>=0.14.0
# Redis
redis>=5.2.0
python-dotenv>=1.0.0
httpx>=0.28.0  # для внешних API запросов
jinja2>=3.1.5  # для шаблонов

# Testing
pytest>=8.3.0
pytest-asyncio>=0.25.0
pytest-cov>=6.0.0

# Security
bcrypt>=4.2.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.4.0

# Monitoring
sentry-sdk[fastapi]>=2.20.0
# Rate limiting
slowapi>=0.1.9
# Prometheus metrics
prometheus-client>=0.21.0
psutil>=6.1.0

# Development dependencies
black>=25.1.0
isort>=6.0.0
flake8>=7.1.0
mypy>=1.15.0
pylint>=3.3.0
radon>=6.0.0
bandit>=1.8.0
safety>=3.2.0

# WebSocket
python-socketio>=5.12.0